Fixing issues with dependency check
