+++ package.json
@@ -17,8 +17,6 @@
   },
   "license": "MIT",
   "devDependencies": {
-    "async": "^0.9.0",
-    "detective": "^4.0.0",
     "grunt": "~0.4.2",
     "grunt-contrib-clean": "~0.5.0",
     "grunt-contrib-jshint": "~0.8.0",
@@ -32,6 +30,6 @@
   "dependencies": {
     "async": "^0.9.0",
     "detective": "^4.0.0",
-    "grunt-dependency-check": "^1.1.0"
+    "dependency-check": "^2.4.0"
   }
 }
